Further restructuring
By now we've cleaned up the security and IANA sections that are expected in all
documents; however the one thing I have left out that might be added later is
some wording around what to do with regards to non-NTP specific IANA registry
information - for example, there is a DHCPv6 option for 5905. Do we define a new
option, or use compatibility?