Rewording requirements (#6)
This fairly large commit largely covers replacing the wording so that the
specification is not be defined, but the requirements of the protocol are. This
also sneaks in a bunch of other changes to normative language that will fix some
other things, that, quite frankly, I can't be bothered to go into length
explaining here.