Various nit fixing
This should bring the document a bit more in line with RFC editor requirements
set in RFC7322, with the notable exception that I am using only a single space
after full-stop of sentence - although Section 3.2 says that:

> When a sentence ended by a period is immediately followed by
> another sentence, there must be two blank spaces after the period.

The reality is based on looking at the raw XML of very recent published RFCs is
that is not enforced with a mixture of single and double spaces being present.
My document is now at least consistent, even if it may be wrong.