Adjusting data centre use case
I disagree with Ulrich and don't see an issue mentioning PTP, and there's more
detail I could go through about why, but given it's informational and an example
not sure spending more time on this is worth it particularly since I don't think
it causes confusion or is misleading in any way.

This should resolve #40.